add musicinfo command to look for music
the command is executable but it's way too unstable and it's best to not use it for awhile. this command was suggested by my friend on discord to recommend me to make a command to look up info about certain music or artist, maybe both. this command would be in development stage for awhile, I'll continue this side gig when i have the time